requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.1.0
redis>=5.0.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
google-cloud-run>=0.10.0
